backend/requirements.txt
# NovaTel AI Support Agent — Backend Dependencies

# ── Web Framework ─────────────────────────────────────────────────────────────
fastapi==0.115.0
uvicorn==0.30.6

# ── LangChain & LangGraph ─────────────────────────────────────────────────────
langchain==0.3.7
langchain-openai==0.2.9
langchain-community==0.3.7
langchain-chroma==0.1.4
langgraph==0.2.50

# ── Vector Store ──────────────────────────────────────────────────────────────
chromadb==0.5.18

# ── OpenAI ───────────────────────────────────────────────────────────────────
openai==1.54.5

# ── Document Loading ──────────────────────────────────────────────────────────
unstructured[docx]==0.16.5
python-docx==1.1.2

# ── Data Validation ───────────────────────────────────────────────────────────
pydantic==2.9.2
pydantic-settings==2.6.0

# ── Utilities ─────────────────────────────────────────────────────────────────
python-dotenv==1.0.1
tiktoken==0.8.0
httpx==0.27.2

# ── Optional: Cross-encoder reranking (no API cost) ───────────────────────────
# sentence-transformers==3.3.1
